mini_projects/kiss.daniel/2. házi feladat/requirements.txt
fastapi
uvicorn[standard]
langgraph
qdrant-client
sentence-transformers
httpx
pydantic
numpy
torch
python-dotenv
